requirements.txt
watchdog>=4.0.0
colorlog>=6.8.0
PyYAML>=6.0.1
Jinja2>=3.1.4
pytest>=8.0.0
pytest-cov>=5.0.0
